Rename src to kb_guardian, move callables to separate scripts folder and fix missing docstrings
