+++ kb_guardian/contradiction_detection.py
@@ -1,5 +1,3 @@
-import argparse
-import os
 from typing import List
 
 from langchain.chains import RetrievalQAWithSourcesChain
@@ -12,39 +10,16 @@ from langchain.schema import Document
 from langchain.vectorstores import VectorStore
 from langchain.vectorstores.base import VectorStoreRetriever
 
-from docs_chunks_conversion import create_document_chunks
-from logger import INFO_LOGGER, log_contradiction_result
-from utils.deployment import get_deployment_llm
-from utils.paths import get_config, get_data_folders, get_vectorstore_paths
-from utils.vectorstore import load_FAISS_vectorstore, save_FAISS_vectorstore
+from kb_guardian.logger import INFO_LOGGER, log_contradiction_result
+from kb_guardian.utils.deployment import get_deployment_llm
+from kb_guardian.utils.paths import get_config
 
 CONFIG = get_config()
-
-(
-    INDEX_PATH,
-    VECTORSTORE_PATH,
-    EXTENDED_INDEX_PATH,
-    EXTENDED_VECTORSTORE_PATH,
-) = get_vectorstore_paths(CONFIG)
-RAW_DATA_FOLDER, EXTENSION_DATA_FOLDER = get_data_folders(CONFIG)
-
-
-def parse_arguments():
-    parser = argparse.ArgumentParser(description="Extend vectorstores")
-
-    parser.add_argument(
-        "--force-extend",
-        dest="force_extend",
-        default=False,
-        help="Force the extension of an existing vectorstore",
-    )
-    args = parser.parse_args()
-    return args
 
 
 def get_detection_chain(retriever: VectorStoreRetriever) -> RetrievalQAWithSourcesChain:
     """
-    Constructs and returns an LLM chain to detect contradictions at ingestion time.
+    Construct and return an LLM chain to detect contradictions at ingestion time.
 
     Args:
         retriever (VectorStoreRetriever): A retriever for the vector store to which you want to add new documents
@@ -81,7 +56,8 @@ def contradiction_detection(
     chunks: List[Document],
 ) -> VectorStore:
     """
-    This method uses an LLM to detect inconsistencies / contradictions between the documents in a vectorstore and the documents that should be added to that vectorstore.
+    Use an LLM to detect inconsistencies / contradictions between the documents in a vectorstore and the documents that should be added to that vectorstore.
+
     New documents that do not contradict the existing documents are added to the vectorstore.
     New documents that are contradicting the existing documents are not added. Instead, a log with information about the detected contradictions will be stored.
 
@@ -115,36 +91,3 @@ def contradiction_detection(
     )
 
     return vectorstore
-
-
-if __name__ == "__main__":
-    args = parse_arguments()
-
-    if not (os.path.isfile(VECTORSTORE_PATH)):
-        INFO_LOGGER.error(
-            f"Vectorstore {VECTORSTORE_PATH} does not exist and can therefore not be extended."  # noqa: E501
-        )
-
-    else:
-        vectorstore = load_FAISS_vectorstore(INDEX_PATH, VECTORSTORE_PATH)
-
-        INFO_LOGGER.info("Splitting documents into chunks...")
-        docs_chunks = create_document_chunks(
-            data_path=EXTENSION_DATA_FOLDER,
-            extension=CONFIG["extension"],
-            chunk_size=CONFIG["chunk_size"],
-            chunk_overlap=CONFIG["chunk_overlap"],
-        )
-
-        if args.force_extend:
-            INFO_LOGGER.info(
-                "Extending the vectorstore by force. No contradiction detection will be performed."  # noqa: E501
-            )
-            vectorstore.add_documents(docs_chunks)
-        else:
-            vectorstore = contradiction_detection(vectorstore, docs_chunks)
-
-        INFO_LOGGER.info("Saving the extended vectorstore...")
-        save_FAISS_vectorstore(
-            vectorstore, EXTENDED_INDEX_PATH, EXTENDED_VECTORSTORE_PATH
-        )
